+++ test/webpack/create-controllers-module.test.js
@@ -15,22 +15,22 @@ describe('createControllersModule', () => {
     describe('empty.json', () => {
         it('must return an empty file', () => {
             const config = require('../fixtures/empty.json');
-            expect(createControllersModule(config)).toEqual('export default {\n};');
+            expect(createControllersModule(config).finalSource).toEqual('export default {\n};');
         });
     });
 
     describe('disabled-controller.json', () => {
         it('must return an empty file', () => {
             const config = require('../fixtures/disabled-controller.json');
-            expect(createControllersModule(config)).toEqual('export default {\n};');
+            expect(createControllersModule(config).finalSource).toEqual('export default {\n};');
         });
     });
 
     describe('disabled-autoimport.json', () => {
         it('must return file with no autoimport', () => {
             const config = require('../fixtures/disabled-autoimport.json');
-            expect(createControllersModule(config)).toEqual(
-                "export default {\n  '@symfony/mock-module/mock': import(/* webpackMode: \"lazy\" */ '@symfony/mock-module/dist/controller.js'),\n};"
+            expect(createControllersModule(config).finalSource).toEqual(
+                "export default {\n  'symfony--mock-module--mock': import(/* webpackMode: \"eager\" */ '@symfony/mock-module/dist/controller.js'),\n};"
             );
         });
     });
@@ -38,17 +38,59 @@ describe('createControllersModule', () => {
     describe('eager-no-autoimport.json', () => {
         it('must return file with no autoimport', () => {
             const config = require('../fixtures/eager-no-autoimport.json');
-            expect(createControllersModule(config)).toEqual(
-                "export default {\n  '@symfony/mock-module/mock': import(/* webpackMode: \"eager\" */ '@symfony/mock-module/dist/controller.js'),\n};"
+            expect(createControllersModule(config).finalSource).toEqual(
+                "export default {\n  'symfony--mock-module--mock': import(/* webpackMode: \"eager\" */ '@symfony/mock-module/dist/controller.js'),\n};"
             );
         });
     });
 
-    describe('lazy-autoimport.json', () => {
-        it('must return a file with the enabled controller', () => {
-            const config = require('../fixtures/lazy-autoimport.json');
-            expect(createControllersModule(config)).toEqual(
-                "import '@symfony/mock-module/dist/style.css';\nexport default {\n  '@symfony/mock-module/mock': import(/* webpackMode: \"lazy\" */ '@symfony/mock-module/dist/controller.js'),\n};"
+    describe('deprecated-webpack-mode.json', () => {
+        it('must return eager mode with deprecation', () => {
+            const config = require('../fixtures/deprecated-webpack-mode.json');
+            const result = createControllersModule(config);
+            expect(result.finalSource).toEqual(
+                "export default {\n  'symfony--mock-module--mock': import(/* webpackMode: \"eager\" */ '@symfony/mock-module/dist/controller.js'),\n};"
+            );
+            expect(result.deprecations).toHaveLength(1);
+        });
+    });
+
+    describe('eager-autoimport.json', () => {
+        it('must return a file with the enabled controller and auto-import', () => {
+            const config = require('../fixtures/eager-autoimport.json');
+            expect(createControllersModule(config).finalSource).toEqual(
+                "import '@symfony/mock-module/dist/style.css';\nexport default {\n  'symfony--mock-module--mock': import(/* webpackMode: \"eager\" */ '@symfony/mock-module/dist/controller.js'),\n};"
+            );
+        });
+    });
+
+    describe('lazy-controller-no-autoimport.json', () => {
+        it('must return a file with a lazy controller', () => {
+            const config = require('../fixtures/lazy-no-autoimport.json');
+            expect(createControllersModule(config).finalSource).toEqual(
+                `
+import { Controller } from 'stimulus';
+export default {
+  'symfony--mock-module--mock': new Promise((resolve, reject) => resolve({ default:
+      (function() {
+          function LazyController(context) {
+              Controller.call(this, context);
+          }
+          LazyController.prototype = Object.create(Controller && Controller.prototype, {
+              constructor: { value: LazyController, writable: true, configurable: true }
+          });
+          Object.setPrototypeOf(LazyController, Controller);
+          LazyController.prototype.initialize = function() {
+              var _this = this;
+              import('@symfony/mock-module/dist/controller.js').then(function(controller) {
+                  _this.application.register(_this.identifier, controller.default);
+              });
+          }
+          return LazyController;
+      })()
+  })),
+};
+                `.trim()
             );
         });
     });